pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "demo"
version = "0.1.0"
description = "Demo project."
authors = [
    { name = "Frédéric Meurou", email = "fmeurou@odolix.fr" },
]
dependencies = [
    "Django>=5.1",
    "djangorestframework~=3.15.0",
    "drf-spectacular~=0.27.0",
    "python-dotenv~=1.0.0",
    "django-filter~=24.0",
    "redis~=5.0.0",
    "unidecode~=1.3.6",
    "black>=24.10.0",
    "pylint>=3.3.1",
    "ruff>=0.7.0",
    "pre-commit>=4.0.1",
]
requires-python = ">=3.10,<3.13"
readme = "README.md"

[tool.pdm.build]
includes = [
    "demo/",
    "demo/**/*.py",
    "demo/**/*.css",
    "demo/**/*.html",
    "demo/**/*.js"
]
excludes = [
    "tests",
]

[project.group.dev.dependencies]
django-extensions = "~=3.2"
black = "~=23.1"
